requirements.txt
alembic==1.15.2
asyncpg==0.30.0
bcrypt==4.3.0
fastapi[standard]==0.115.12
sqlalchemy==2.0.40
passlib==1.7.4
psycopg2==2.9.10
pydantic_settings==2.9.1
PyJWT==2.10.1
redis==6.1.0
